Restructure Providers
